Turkish Translation for B41 and B42
Added Turkish Translation for B41 and B42 + Added EN folder to B42 Translations.

I used Windows-1254 for B41 and UTF-8 for B42. Unfortunately UTF-8 for Turkish does not work in version B41 so I had to use Windows-1254.

I apologize for any mistakes. When I tried it on my local computer, it worked fine. If there is an Encode error I can fix it with only 1 more update.